Improved: diff logic for filter and sort options(#146)
When having same fieldName in sort and filter options and adding/removing that field in both filter and sort, then always the sort option is honored and the filter option is ignored, as the filter field key gets overridden by the sort key inside the object.
Updated the keys after finding the diff to include filter and sort string in the respective options, to handle the override property issue.